Commit demonstrating more in depth commit message
More detailed explanatore text, if necessary. Wrap it at about 72 chars
or so. In some contexts, the first line is treated as the subject of the
commit and the rest as the body. The blank line seperating the two is
critical; various tools like `log`, `shortlog`, and `rebase` can get
confused if you run the two together.

Further paragraphs come after blank lines.

  - Bullet points are okay too

  - typically, a hyphen or asterisk is used for a bullet, preceeded by a
    single space, with blank lines in between, but convention varies
    here

If you use an issue tracker, put reference to them at the bottom,
like this:

Resolves: #69
See also: #420